+++ tests/Unit/Serializer/AuditLogMessageSerializerTest.php
@@ -11,6 +11,7 @@ use Rcsofttech\AuditTrailBundle\AuditTrailBundle;
 use Rcsofttech\AuditTrailBundle\Message\AuditLogMessage;
 use Rcsofttech\AuditTrailBundle\Message\Stamp\ApiKeyStamp;
 use Rcsofttech\AuditTrailBundle\Message\Stamp\SignatureStamp;
+use Rcsofttech\AuditTrailBundle\Serializer\AuditLogMessagePayloadEncoder;
 use Rcsofttech\AuditTrailBundle\Serializer\AuditLogMessageSerializer;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
@@ -86,5 +87,27 @@ final class AuditLogMessageSerializerTest extends TestCase
         $this->expectExceptionMessage('Decoding is not supported');
 
         $this->serializer->decode(['body' => '{}']);
+    }
+
+    public function testEncodeUsesCanonicalJsonBody(): void
+    {
+        $message = new AuditLogMessage(
+            'TestEntity',
+            '123',
+            'create',
+            null,
+            ['url' => 'https://a/b', 'label' => 'Resume', 'title' => 'Resume'],
+            ['url', 'label', 'title'],
+            null,
+            null,
+            null,
+            null,
+            null,
+            '2024-01-01T00:00:00+00:00',
+        );
+
+        $encoded = $this->serializer->encode(new Envelope($message));
+
+        self::assertSame(new AuditLogMessagePayloadEncoder()->encode($message), $encoded['body']);
     }
 }